Move smolagents and langchain to mixins
